+++ tests/controllers/body/test1/index.js
@@ -1,0 +1,10 @@
+'use strict';
+var body = {};
+
+module.exports = body;
+
+body.get = function*() {
+    this.body = {
+        "hello": "world!"
+    }
+};