+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1
+- Added checks to ensure TLS cert and key exist (closes #7)
+- Fixed error in controllers not displaying to console (closes #8)
+- Added ability to override default json structure by setting this.body directly
+
 # 1.1.0
 - Changed required field within JWT to 'sub' from 'userId'
 - Added ability to add your own RAML trait names that should be authenticated by JWT